Collision (not working) + double -> float

+++ include/debug.hpp
@@ -18,7 +18,7 @@ void debug_init();
 #define _debug_log(args...) fprintf(debug_file, args)
 #define debug_log(args...) {_debug_log(args); fflush(debug_file);}
 
-void debug_log_obj(double n);
+void debug_log_obj(float n);
 void debug_log_obj(int n);
 void debug_log_obj(vec3 vec);
 void debug_log_obj(triangle3 tri);
